+++ mvc/src/main/java/com/interface21/webmvc/servlet/adapter/HandlerAdapterRegistry.java
@@ -1,9 +1,6 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.adapter;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.adapter.ControllerHandlerAdapter;
-import com.interface21.webmvc.servlet.adapter.HandlerAdapter;
-import com.interface21.webmvc.servlet.adapter.HandlerExecutionHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
